clearFileCache resolvers moved to pipelines and deployments
