+++ app/mainapp/graphql/private/resolvers/code_editor.resolvers.go
@@ -647,50 +647,6 @@ func (r *mutationResolver) UpdateCodePackages(ctx context.Context, workerGroup s
 	return "Success", nil
 }
 
-func (r *mutationResolver) ClearFileCachePipeline(ctx context.Context, environmentID string, pipelineID string) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Subject: "user", SubjectID: currentUser, Resource: "admin_platform", ResourceID: platformID, Access: "write", EnvironmentID: "d_platform"},
-		{Subject: "user", SubjectID: currentUser, Resource: "admin_environment", ResourceID: environmentID, Access: "write", EnvironmentID: environmentID},
-		{Subject: "user", SubjectID: currentUser, Resource: "environment_edit_all_pipelines", ResourceID: platformID, Access: "write", EnvironmentID: environmentID},
-		{Subject: "user", SubjectID: currentUser, Resource: "specific_pipeline", ResourceID: pipelineID, Access: "write", EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	return "Success", nil
-	// panic(fmt.Errorf("not implemented"))
-}
-
-func (r *mutationResolver) ClearFileCacheDeployment(ctx context.Context, environmentID string, deploymentID string) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Subject: "user", SubjectID: currentUser, Resource: "admin_platform", ResourceID: platformID, Access: "write", EnvironmentID: "d_platform"},
-		{Subject: "user", SubjectID: currentUser, Resource: "admin_environment", ResourceID: environmentID, Access: "write", EnvironmentID: environmentID},
-		{Subject: "user", SubjectID: currentUser, Resource: "environment_edit_all_pipelines", ResourceID: platformID, Access: "write", EnvironmentID: environmentID},
-		{Subject: "user", SubjectID: currentUser, Resource: "specific_pipeline", ResourceID: deploymentID, Access: "write", EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	return "Success", nil
-	// panic(fmt.Errorf("not implemented"))
-}
-
 func (r *queryResolver) FilesNode(ctx context.Context, environmentID string, nodeID string, pipelineID string) (*privategraphql.CodeTree, error) {
 	currentUser := ctx.Value("currentUser").(string)
 	platformID := ctx.Value("platformID").(string)
